requirements.txt
numpy>=1.24.0
torch>=2.0.0
Pillow>=9.0.0 